modify commands
